+++ python/ray/tests/test_advanced_2.py
@@ -520,6 +520,21 @@ def test_zero_capacity_deletion_semantics(shutdown_only):
     assert cluster_resources == {}
 
 
+def test_ray_get_timeout_zero(monkeypatch):
+    # Check that ray.get(timeout=0) raises warnings on change of behavior.
+    # Removed when https://github.com/ray-project/ray/issues/28465 is resolved.
+    with pytest.warns(UserWarning):
+        ray.get(ray.put(1), timeout=0)
+
+    with monkeypatch.context() as m:
+        m.setenv("RAY_WARN_RAY_GET_TIMEOUT_ZERO", "0")
+        import warnings
+
+        with warnings.catch_warnings():
+            warnings.simplefilter("error")
+            ray.get(ray.put(1), timeout=0)
+
+
 if __name__ == "__main__":
     if os.environ.get("PARALLEL_CI"):
         sys.exit(pytest.main(["-n", "auto", "--boxed", "-vs", __file__]))
